pyproject.toml
[project]
name = "python-circuitpython"
version = "1.0.0-beta"
requires-python = "==3.14.1"
dependencies = [
    "adafruit-blinka>=8.0.0",
    "adafruit-circuitpython-dht>=3.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "aiomqtt>=2.0.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "black[d]>=24.10.0",
    "mypy>=1.14.0",
    "poethepoet>=0.32.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "pip-audit>=2.6.1",
    "pytest-cov>=6.0.0",
    "deptry>=0.23.0",
    "pip>=25.1.1",
    "paramiko>=3.0.0",
    "pytest-xdist>=3.8.0",
    "filelock>=3.0.0",
    "testcontainers>=4.0.0",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_also = [
    # Don't complain about abstract methods that aren't implemented
    "@(abc\\.)?abstractmethod",
    # Don't complain about type checking blocks
    "if TYPE_CHECKING:",
    # Don't complain about ellipsis in protocols
    "\\.\\.\\.",
]
# Show which lines are missing coverage
show_missing = true
# Skip files with 100% coverage
skip_covered = false

[tool.poe.tasks]
depcheck = "deptry ."
format = "black ."
lint = "ruff check"
typecheck = "mypy  --install-types --non-interactive  --explicit-package-bases ."
audit-src = "bandit -r src/* -s B101"
audit-packages = "pip-audit --ignore-vuln GHSA-4xh5-x5gv-qwph"
security = ["audit-src", "audit-packages"]
checks = ["depcheck", "format", "lint", "typecheck", "security"]
unit = "pytest -vv src/*"
sensors-test = "pytest -d --tx ssh=${PI_HOST:-pi@raspberrypi.local}//python=python3 --rsyncdir src --rsyncdir tests tests/test_sensors.py -p no:warnings"
integration = "pytest -vv tests/test_integration.py"
healthcheck = "bash -c 'ssh -o ConnectTimeout=10 -o BatchMode=yes ${PI_HOST:-pi@raspberrypi.local} echo NETWORK_READY'"
test = ["unit", "healthcheck", "sensors-test", "integration"]
cover = "pytest -vv --cov --cov-branch --cov-report=term-missing src/"
review = "claude 'run code-reviewer agent against diff for simplicity and elegance'"
commit = "bash -c 'uv run poe checks && uv run poe test && uv run poe review && git add -A && git cz && git push'"
